Updated expat to 2.2.0, added _NO_THERMORAW definition to exclude thermo raw drivers and support from Windows builds.

+++ include/mzParser.h
@@ -342,7 +342,7 @@ public:
 	//Accessors
 	vector<TimeIntensityPair>&	getData();
 	int													getIDString(char* str);
-	unsigned int								size();
+	size_t      								size();
 
 protected:
 
@@ -1395,7 +1395,7 @@ public:
 
 	ChromatogramPtr		chromatogram(int index, bool binaryData = false);
 	bool							get();
-	unsigned int			size();
+	size_t      			size();
 
 	vector<cindex>*			vChromatIndex;
   #ifdef MZP_MZ5
